+++ api-js/src/validators/__tests__/cleanse-input.test.js
@@ -1,4 +1,4 @@
-const { cleanseInput } = require('../validators')
+const { cleanseInput } = require('..')
 
 describe('given an input validate it', () => {
   describe('string contains symbols', () => {
